+++ .travis.yml
@@ -11,7 +11,7 @@ install:
   - gem install coveralls-lcov
 
 script:
-  - make test
+  - make c
 
 after_success:
   - coveralls-lcov /home/travis/build/tecnickcom/binsearch/target/test/coverage/binsearch.info
