+++ README.md
@@ -37,9 +37,8 @@ xxd -r -p hexfile.txt > binaryfile.bin
 
 ## Getting Started
 
-This application is written in C language and it is C99 compatible.
+The reference code of this application is written in C language and includes wrappers for GO and Python.
 
-This project can be built using CMake.
 A wrapper Makefile is available to allows building the project in a Linux-compatible system with simple commands.  
 All the artifacts and reports produced using this Makefile are stored in the *target* folder.  
 
@@ -48,24 +47,4 @@ To see all available options:
 make help
 ```
 
-To execute all the default test builds and generate reports in the current environment:
-```
-make test
-```
-
-To format the code (please use this command before submitting any pull request):
-```
-make format
-```
-
-To build a library:
-```
-make build
-```
-Set *BS_BUILD_SHARED_LIB* in config.mk to build a static or shared library.
-
-
-## Python and GO Libraires
-
-Implementation in Python, GO and CGO are available.
-use the command ```make all``` to build a test all the implementations.
+use the command ```make all``` to build and test all the implementations.
