+++ c/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required (VERSION 3.1)
 project (binsearch)
 
-set( CMAKE_VERBOSE_MAKEFILE off )
+set( CMAKE_VERBOSE_MAKEFILE on )
 
 add_definitions(-DCMAKE)
 
@@ -9,23 +9,25 @@ add_definitions(-DCMAKE)
 enable_testing ()
 
 # Extract version from the VERSION file
-file(STRINGS VERSION VERSION_FILE_CONTENT)
+file(STRINGS ../VERSION VERSION_FILE_CONTENT)
+file(STRINGS ../RELEASE PROJECT_RELEASE)
 string(REPLACE "." ";" VERSION_FILE_PARTS ${VERSION_FILE_CONTENT})
-list(GET VERSION_FILE_PARTS 0 BS_VERSION_MAJOR)
-list(GET VERSION_FILE_PARTS 1 BS_VERSION_MINOR)
-list(GET VERSION_FILE_PARTS 2 BS_VERSION_PATCH)
-set(PROJECT_VERSION "${BS_VERSION_MAJOR}.${BS_VERSION_MINOR}.${BS_VERSION_PATCH}")
+list(GET VERSION_FILE_PARTS 0 PROJECT_VERSION_MAJOR)
+list(GET VERSION_FILE_PARTS 1 PROJECT_VERSION_MINOR)
+list(GET VERSION_FILE_PARTS 2 PROJECT_VERSION_PATCH)
+set(PROJECT_VERSION "${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}.${PROJECT_VERSION_PATCH}-${PROJECT_RELEASE}")
+
 message(STATUS "PROJECT_VERSION='${PROJECT_VERSION}'")
 message(STATUS "Generator - ${CMAKE_GENERATOR}")
 
-option (BUILD_DOXYGEN "Build Doxygen" OFF)
-option (BUILD_SHARED_LIB "Build a shared library" OFF)
+option(BUILD_DOXYGEN "Build Doxygen" OFF)
+option(BUILD_SHARED_LIB "Build a shared library" ON)
 
 if(CMAKE_COMPILER_IS_GNUCC)
     message(STATUS "Build type: ${CMAKE_BUILD_TYPE}")
     execute_process(COMMAND ${CMAKE_C_COMPILER} -dumpversion
                     OUTPUT_VARIABLE GCC_VERSION)
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -pedantic -std=c99 -Wall -Wextra -Wno-strict-prototypes -Wunused-value -Wcast-align -Wundef -Wformat-security")
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -D VERSION='\"${PROJECT_VERSION}\"' -s -pedantic -std=c99 -Wall -Wextra -Wno-strict-prototypes -Wunused-value -Wcast-align -Wundef -Wformat-security")
 
     if (GCC_VERSION VERSION_GREATER 4.8 OR GCC_VERSION VERSION_EQUAL 4.8)
         set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wshadow")
@@ -67,6 +69,21 @@ if(${CMAKE_SYSTEM_NAME} MATCHES "Windows")
    set(OS "Windows")
 endif(${CMAKE_SYSTEM_NAME} MATCHES "Windows")
 
+# Common package settings
+set(CPACK_PACKAGE_VENDOR "Tecnick.com")
+set(CPACK_PACKAGE_CONTACT "info@tecnick.com")
+set(CPACK_PACKAGE_VERSION_MAJOR "${PROJECT_VERSION_MAJOR}")
+set(CPACK_PACKAGE_VERSION_MINOR "${PROJECT_VERSION_MINOR}")
+set(CPACK_PACKAGE_VERSION_PATCH "${PROJECT_VERSION_PATCH}")
+set(CPACK_PACKAGE_VERSION "${PROJECT_VERSION}")
+set(CPACK_PACKAGE_CHECKSUM "SHA512")
+set(CPACK_PACKAGE_DESCRIPTION_FILE "${CMAKE_CURRENT_SOURCE_DIR}/../README.md")
+set(CPACK_RESOURCE_FILE_README "${CMAKE_CURRENT_SOURCE_DIR}/../README.md")
+set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/../LICENSE")
+set(CPACK_GENERATOR "DEB;RPM;TGZ")
+set(CPACK_COMPONENTS_IGNORE_GROUPS TRUE)
+set(CPACK_DEB_COMPONENT_INSTALL TRUE)
+
 # Add subdirectories
 add_subdirectory(src)
 add_subdirectory(test)
@@ -74,7 +91,7 @@ add_subdirectory(test)
 # Build Documentation
 find_package(Doxygen QUIET)
 if(DOXYGEN_FOUND)
-    add_subdirectory (doc)
+    add_subdirectory(doc)
 else(DOXYGEN_FOUND)
     message(STATUS "Doxygen not found. Documentation will not be built.")
 endif(DOXYGEN_FOUND)
