+++ cgo/src/binsearch_test.go
@@ -601,7 +601,7 @@ func TestFindLastUint64Sub(t *testing.T) {
 }
 
 func TestClose(t *testing.T) {
-	lmf, err := MmapBinFile("../../test/data/test_data_b.bin")
+	lmf, err := MmapBinFile("../../c/test/data/test_data_b.bin")
 	if err != nil {
 		t.Errorf("Unexpected error: %v", err)
 	}
