+++ conda/channels.mk
@@ -1,0 +1,4 @@
+# This file sets the default build of the library and is run by typing "make"
+
+# Conda channels
+ARTIFACTORY_CONDA_CHANNELS=-c conda-forge -c https://repo.continuum.io/pkgs/main -c https://repo.continuum.io/pkgs/free
