+++ conda/python/meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: binsearch
-  version: 4.0.0.1
+  version: 4.0.0
 
 source:
-  path: ..
+  path: ../..
 
 build:
   number: 1
